refactor: Refactor according to yael's recommendations
