+++ src/services/UsersService.js
@@ -1,15 +1,11 @@
 import axios from 'axios';
 import config from '../appConf';
 
-const headers = {
-  Accept: 'application/json',
-};
-
 class UsersService {
   static async searchUsers(searchValue) {
     const { data } = await axios.get(`${config.uri.api_gateway_uri}api/users`,
       { params: { partialName: searchValue } },
-      { ...headers });
+      { Accept: 'application/json' });
     return data;
   }
 }
